requirements.txt
duckdb==1.0.0
python-dotenv
deltalake==0.18.2
pyarrow==17.0.0
pandas==2.2.2
boto3==1.34.144
